Fix references that weren't updated during the move
