+++ Source/SelfService/Web/integrations/connection/connectionDetails/configuration/setup/M3ConfigurationForm.tsx
@@ -3,17 +3,14 @@
 
 import React, { useImperativeHandle, useState, useCallback, useEffect, useMemo } from 'react';
 
-import { useSnackbar } from 'notistack';
-import { useQueryClient } from '@tanstack/react-query';
 import { SubmitHandler } from 'react-hook-form';
 
 import { Form, FormRef } from '@dolittle/design-system';
 
-import { CACHE_KEYS } from '../../../../apis/integrations/CacheKeys';
-import { ConnectionModel, ConnectionModelResult, IonConfigRequest, IonConfigurationResult, M3BasicAuthConfigRequest, M3BasicAuthConfigurationResult, MdpConfigurationResult, StringResult } from '../../../../apis/integrations/generated';
-import { useConnectionsIdNamePost } from '../../../../apis/integrations/connectionsApi.hooks';
-import { useConnectionsIdDeployCloudPost, useConnectionsIdDeployOnPremisesPost } from '../../../../apis/integrations/deploymentApi.hooks';
-import { useConnectionsIdConfigurationMdpPost, useConnectionsIdConfigurationIonPost, useConnectionsIdConfigurationBasicPost } from '../../../../apis/integrations/connectionConfigurationApi.hooks';
+import { ConnectionModel, ConnectionModelResult, IonConfigRequest, IonConfigurationResult, M3BasicAuthConfigRequest, M3BasicAuthConfigurationResult, MdpConfigurationResult, StringResult } from '../../../../../apis/integrations/generated';
+import { useConnectionsIdNamePost } from '../../../../../apis/integrations/connectionsApi.hooks';
+import { useConnectionsIdDeployCloudPost, useConnectionsIdDeployOnPremisesPost } from '../../../../../apis/integrations/deploymentApi.hooks';
+import { useConnectionsIdConfigurationMdpPost, useConnectionsIdConfigurationIonPost, useConnectionsIdConfigurationBasicPost } from '../../../../../apis/integrations/connectionConfigurationApi.hooks';
 
 //TODO: Can this be replaced with using `watch` from react-hook-form?
 const useForceSubscribeToIonConfigurationStateChanges = (currentForm: FormRef<M3ConnectionParameters> | undefined) => {
